requirements.txt
openai
python-telegram-bot==20.6
feedparser
apscheduler
